requirements.txt
awsebcli==3.14.2
Babel==2.6.0
backcall==0.1.0
beautifulsoup4==4.6.0
botocore==1.10.58
cement==2.8.2
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
decorator==4.3.0
Django==2.0.5
django-cors-headers==2.2.1
django-db-file-storage==0.5.2
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-mysql==2.2.2
django-phonenumber-field==2.0.0
django-webpack-loader==0.6.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docutils==0.14
et-xmlfile==1.0.1
generic==0.3.1
idna==2.7
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
jdcal==1.4
jedi==0.12.0
jmespath==0.9.3
jsonfield==2.0.2
lxml==4.2.2
Markdown==2.6.11
mysqlclient==1.3.12
openpyxl==2.5.3
parso==0.2.1
pathspec==0.5.5
phonenumberslite==8.9.9
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
py==1.5.2
Pygments==2.2.0
PyJWT==1.6.4
pytest==3.2.5
python-dateutil==2.7.3
python-dotenv==0.8.2
pytz==2018.4
PyYAML==3.12
requests==2.19.0
semantic-version==2.5.0
simplegeneric==0.8.1
six==1.11.0
South==1.0.2
sqlparse==0.2.4
tabulate==0.7.5
termcolor==1.1.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
